Move viewport resize observers
